+++ src/main/java/chess/dto/response/ChessGameDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.response;
 
 import chess.domain.ChessGame;
 import chess.domain.ChessMap;
