+++ src/main/java/chess/dto/response/StatusDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.response;
 
 import chess.domain.GameResult;
