+++ src/main/java/chess/service/ChessGameService.java
@@ -14,12 +14,12 @@ import chess.domain.piece.State;
 import chess.domain.player.Player;
 import chess.domain.player.Team;
 import chess.domain.position.Position;
-import chess.dto.ChessGameDto;
-import chess.dto.ChessGameInfoDto;
-import chess.dto.CreateGameDto;
-import chess.dto.DeleteGameDto;
 import chess.dto.PieceDto;
-import chess.dto.StatusDto;
+import chess.dto.request.CreateGameDto;
+import chess.dto.request.DeleteGameDto;
+import chess.dto.response.ChessGameDto;
+import chess.dto.response.ChessGameInfoDto;
+import chess.dto.response.StatusDto;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.stereotype.Service;
